Fixes for Empty Token + Test case set up but commented out + Deleting bad logs
Test is commented out because it doesn't currently fit in the testing framework and will need to be revisited later. The main issue that it doesn't fit the framework is because the test case we want to have fails right away with an error message and the current framework is a request/response type of framework, so since theres no response the framework will just hang until timeout.